requirements.txt
gradio
requests
python-dotenv
pillow
librosa